LayoutTests/imported/w3c/web-platform-tests/css/css-pseudo/marker-default-styles-expected.txt


FAIL Computed value of 'unicode-bidi' for outside symbol assert_equals: expected "isolate" but got "normal"
FAIL Computed value of 'font-variant-numeric' for outside symbol assert_equals: expected "tabular-nums" but got "normal"
FAIL Computed value of 'unicode-bidi' for outside decimal assert_equals: expected "isolate" but got "normal"
FAIL Computed value of 'font-variant-numeric' for outside decimal assert_equals: expected "tabular-nums" but got "normal"
FAIL Computed value of 'unicode-bidi' for outside string assert_equals: expected "isolate" but got "normal"
FAIL Computed value of 'font-variant-numeric' for outside string assert_equals: expected "tabular-nums" but got "normal"
FAIL Computed value of 'unicode-bidi' for outside marker assert_equals: expected "isolate" but got "normal"
PASS Computed value of 'font-variant-numeric' for outside marker 
FAIL Computed value of 'unicode-bidi' for inside symbol assert_equals: expected "isolate" but got "normal"
FAIL Computed value of 'font-variant-numeric' for inside symbol assert_equals: expected "tabular-nums" but got "normal"
FAIL Computed value of 'unicode-bidi' for inside decimal assert_equals: expected "isolate" but got "normal"
FAIL Computed value of 'font-variant-numeric' for inside decimal assert_equals: expected "tabular-nums" but got "normal"
FAIL Computed value of 'unicode-bidi' for inside string assert_equals: expected "isolate" but got "normal"
FAIL Computed value of 'font-variant-numeric' for inside string assert_equals: expected "tabular-nums" but got "normal"
FAIL Computed value of 'unicode-bidi' for inside marker assert_equals: expected "isolate" but got "normal"
PASS Computed value of 'font-variant-numeric' for inside marker 
outside symbol
outside decimal
outside string
outside marker
inside symbol
inside decimal
inside string
inside marker
